requirements.txt
Pillow==9.0.1
progressbar2==4.0.0
python-utils==3.1.0
